file structure refactored + helpers and memo optimisation
