+++ src/styles/StopWatchStyles.ts
