--- styles/colors.ts
@@ -1,6 +1,0 @@
-const darkBlue = "#0D1F2D";
-const lightBlue = "#1B2E4B";
-const lighterBlue = "#253B57";
-const accentColor = "#4BA8FF";
-
-export { darkBlue, lightBlue, lighterBlue, accentColor };
